requirements.txt
gTTS==2.5.1
openai_whisper==20231117
PyAudio==0.2.14
streamlit==1.36.0
numpy<2
torch
pyttsx3
pydub